backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "PetalClone - AI Website Cloner Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "anthropic>=0.40.0",
    "openai>=1.52.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "aiofiles>=24.1.0",
    "playwright>=1.48.0",
]
